* '4.6' of scm.dev.nokia.troll.no:qt/oslo-staging-2:
  Fixed some global QIcon/QPixmap instances that leaked handles on X11.
  Speed up QListView test
  QListView: fix crash when hiding many of the lasts item in a QListView
  Fixed warnings and crash when painting graphics effects outside scene.
  Stabilize QLineEdit test on X11
  (sqlite) Allow shared cache mode
  Make generate uid3 (symbian) work on 64 bit host platform.
  Updated the docs for QPainter::begin/endNativePainting()
  Compile fix for network benchmarks.
  Add a pixmap modification hook to blur pixmap filter cache
  Delete benchmark examples (qtestlib-simple and qtwidgets).
  Restructure tests/benchmarks directory.
  Fixed QImagReader::setAutoDetectImageFormat() to work with plugins.
  QLineEdit: regression: read-only line edits would eat shortcuts.
  QGraphicsItem: Do not crash at exit if there is static QGraphicsItem.
  Make QTextCodec reentrant.
  Fixed bug where GL widget was not fully updated on Vista.

+++ qmake/generators/symbian/initprojectdeploy_symbian.cpp
@@ -207,7 +207,7 @@ QString generate_uid(const QString& target)
         return tmp;
     }
 
-    unsigned long hash = 5381;
+    quint32 hash = 5381;
     int c;
 
     for (int i = 0; i < target.size(); ++i) {
